+++ packages/databricks-vscode/src/configuration/auth/DatabricksCliCheck.ts
@@ -84,6 +84,7 @@ export class DatabricksCliCheck implements Disposable {
                 host: this.authProvider.host.toString(),
                 authType: "databricks-cli",
                 databricksCliPath: this.authProvider.cliPath,
+                profile: this.authProvider.profile,
             },
             {
                 product: "databricks-vscode",
@@ -105,9 +106,16 @@ export class DatabricksCliCheck implements Disposable {
     private async login(cancellationToken?: CancellationToken): Promise<void> {
         try {
             const host = this.authProvider.host.toString().replace(/\/+$/, "");
+            const profile = this.authProvider.profile;
+            const args = ["auth", "login"];
+            if (profile) {
+                args.push("--profile", profile);
+            } else {
+                args.push("--host", host);
+            }
             await execFile(
                 this.authProvider.cliPath,
-                ["auth", "login", "--host", host],
+                args,
                 {},
                 cancellationToken
             );